requirements.txt
dj-database-url==0.4.1
Django==1.10.1
djangorestframework==3.4.6
gunicorn==19.6.0
pkg-resources==0.0.0
psycopg2==2.6.2
whitenoise==3.2.1
